+++ gates/generate-showcase.mjs
@@ -133,6 +133,7 @@ function showcase() {
         <link rel="stylesheet" href="../css/components.css" />
         <link rel="stylesheet" href="../css/cyberpunk-register.css" />
         <link rel="stylesheet" href="../css/retro-register.css" />
+        <link rel="stylesheet" href="../css/synthwave-register.css" />
         <style>
 ${STYLE}        </style>
     </head>
@@ -230,6 +231,7 @@ function fixture(theme) {
         <link rel="stylesheet" href="../../css/components.css" />
         <link rel="stylesheet" href="../../css/cyberpunk-register.css" />
         <link rel="stylesheet" href="../../css/retro-register.css" />
+        <link rel="stylesheet" href="../../css/synthwave-register.css" />
         <style>
 ${STYLE}        </style>
     </head>